pyproject.toml
[tool.poetry]
name = "binance-api"
version = "0.1.0"
description = ""
authors = ["nbezmaternykh <nbezmaternykh@7tech.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
binance-connector = "^3.5.1"
pydantic-settings = "^2.1.0"
pika = "^1.3.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
python-telegram-bot = "^20.7"
asyncpg = "^0.29.0"
aio-pika = "^9.4.0"
schedule = "^1.2.1"
apscheduler = "^3.10.4"
pydantic = "2.4.1"
aiogram = "^3.3.0"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
